Redirect STDERR from help to /dev/null
On libreoffice > 4, the -h flag attempts to open the UI to display the
help, in addition to writing it to stdout.

This means that on Linux without X11 running, the output ends up looking like:
    Failed to open display
    LibreOffice 4.1.1.2 410m0(Build:2)

This behavior can be avoided by using the `--headless` flag, but that
flag isn't supported in older version of OpenOffice, causing an error
there.

Since Docsplit only looks at the first line of output it doesn't find
the version string, causing it to assume it should use OpenOffice in the
extract method. It then attempts to use the JOD convertor, which will
fail on recent versions of LibreOffice.

I believe this is the root cause of issue #71 and should correct it.